Refactor caches
Refactor caches into a somewhat better composed layout where there is
a FooService and a FooCache. Application code only interacts with the
FooService, which uses (or may not use) a cache under the hood.
