+++ scripts/search.ts
@@ -5,7 +5,7 @@ import * as redis from 'redis';
 
 import { config } from '../src/config/config';
 import { detect } from '../src/core/detector';
-import { init as initPool } from '../src/core/pools';
+import init from '../src/services/init';
 
 const argv = yargs
     .command('search <address>', 'search for sandwiches', {})
@@ -60,7 +60,7 @@ const log = winston.createLogger({
     const web3 = new Web3(
         argv.web3_url ? argv.web3_url : process.env.WEB3_PROVIDER_URI,
     );
-    initPool(web3, log, redis.createClient());
+    await init(web3, log, redis.createClient());
 
     const from = argv.from;
     const to = argv.to == 'latest' ? await web3.eth.getBlockNumber() : argv.to;
