+++ src/core/sandwich.ts
@@ -3,8 +3,8 @@ import winston from 'winston';
 import { utils, BigNumber } from 'ethers';
 
 import { getSwaps, SwapLog, SwapDir } from './swaps';
-import { Pool, PoolCache } from './pools';
-import { BlockCache } from './blocks';
+import { Pool, PoolService } from '../services/pools';
+import { BlockService } from '../services/blocks';
 
 // temp for CLI... eventually this should just return sandwiches as it
 // finds them (EventEmitter?) and let the caller do what they want,
@@ -28,12 +28,12 @@ interface SwapInfo {
 }
 
 async function SwapInfoFromLog(log: SwapLog): Promise<SwapInfo> {
-    const pool = await PoolCache.lookupOrEnter(log.address);
+    const pool = await PoolService.lookup(log.address);
     if (pool === null) {
         throw new Error('null pool');
     }
 
-    const block = await BlockCache.lookupOrEnter(log.blockNumber);
+    const block = await BlockService.lookup(log.blockNumber);
     const ts =
         typeof block.timestamp === 'string'
             ? block.timestamp
@@ -124,7 +124,7 @@ export async function findSandwich(
         }
 
         const close = closes[0];
-        const pool = await PoolCache.lookupOrEnter(open.address);
+        const pool = await PoolService.lookup(open.address);
         if (pool === null) {
             throw new Error('null pool');
         }
