+++ src/services/blockcache.ts
@@ -1,0 +1,32 @@
+import * as redis from 'redis';
+import { Block } from './blocks';
+import winston from 'winston';
+
+import { RedisClientAsync } from '../redis';
+
+export class BlockCache {
+    static client: RedisClientAsync;
+    static logger: winston.Logger;
+    static init(logger: winston.Logger, redis: redis.RedisClient): void {
+        BlockCache.client = new RedisClientAsync(logger, redis);
+        BlockCache.logger = logger;
+    }
+    static async lookup(bn: number): Promise<Block | null> {
+        const key = BlockCache.key(bn);
+        const block = await BlockCache.client.get(key);
+        if (block == null) {
+            return null;
+        }
+        return JSON.parse(block);
+    }
+    static key(bn: number): string {
+        return `block:${bn}`;
+    }
+
+    static async cache(bn: number, block: Block): Promise<string> {
+        return await BlockCache.client.set(
+            BlockCache.key(bn),
+            JSON.stringify(block),
+        );
+    }
+}
